requirements_minimal.txt
# Minimal requirements for Streamlit Cloud
# Use this if main requirements.txt fails

streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.4
numpy==1.24.3
